Remove Winkki street note updater
