+++ pom.xml
@@ -671,11 +671,6 @@
         </dependency>
         <dependency>
             <groupId>org.geotools</groupId>
-            <artifactId>gt-wfs-ng</artifactId>
-            <version>${geotools.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>org.geotools</groupId>
             <artifactId>gt-geojson-core</artifactId>
             <version>${geotools.version}</version>
         </dependency>
